account-creator/requirements.txt
selenium==4.22.0
psycopg-binary
python-dotenv
requests